feat([sc-5936]): style the iframe holder
