+++ apps/nfid-demo/src/pages/authentication-iframe/components/iframe-wrapper.tsx
@@ -1,8 +1,6 @@
 import clsx from "clsx"
 import React from "react"
 
-import { Card } from "./iframe-card"
-
 interface IFrameWrapperProps extends React.HTMLAttributes<HTMLDivElement> {
   fixedHeight?: boolean
 }
@@ -10,20 +8,18 @@ interface IFrameWrapperProps extends React.HTMLAttributes<HTMLDivElement> {
 export const IFrameWrapper: React.FC<IFrameWrapperProps> = ({
   children,
   className,
-  fixedHeight = true,
 }) => {
   return (
-    <Card
+    <div
       id="iframe-wrapper-card"
       className={clsx(
-        "bg-white shadow-iframe max-w-screen rounded-t-xl md:rounded-xl w-full md:w-[380px] duration-200 ease-in-out overflow-hidden",
+        "bg-white rounded-b-xl md:rounded-b-xl w-full md:w-[380px] duration-200 ease-in-out overflow-hidden",
         "flex flex-col",
-        "fixed bottom-0 right-0 md:top-[18px] md:right-7",
+        "h-[480px] w-[360px]",
         className,
       )}
-      style={{ height: fixedHeight ? 190 : "fit-content" }}
     >
       {children}
-    </Card>
+    </div>
   )
 }
